Dijkstra's algorithm
Dijkstra's algorithm, conceived by Dutch computer scientist Edsger
Dijkstra in 1956 and published in 1959,[1][2] is a graph search
algorithm that solves the single-source shortest path problem for a
graph with non-negative edge path costs, producing a shortest path
tree. This algorithm is often used in routing and as a subroutine in
other graph algorithms.